Module Uspf_unix

Unix layer of uSPF.

uSPF is a standalone library which requires some specialisations like the scheduler and the DNS stack used to get DNS records. This module specialises uSPF with the module Unix and ocaml-dns. check does the SPF verification from a given Uspf.ctx and extract_received_spf extracts the Received-SPF field from a given email.

For more details about uSPF and how to use it, please take a look on Uspf.

val check : ?nameservers:(Dns.proto * (Ipaddr.t * int) list) -> timeout:int64 -> Uspf.ctx -> (Uspf.res, [> `Msg of string ]) Stdlib.result

val extract_received_spf : ?newline:Uspf.newline -> Stdlib.in_channel -> (Uspf.extracted, [> `Msg of string ]) Stdlib.result